Refactor keymap
I think the reason my keymap had to be in matrix order is because I put the key definitions in info.json into matrix order. Duh. I rearranged the json file to be in logical order, and also rearranged the keymap file to be in logical order and column formatted in a way that should make it a lot easier to update. I confirmed it compiles but I haven't actually tested it yet.