Fix my keyboard
Broke my dactyl manuform by getting shocked on it. Had to change a pin, I think the pin driver broke. Left some debug enable code, commented out, for next time. Also changed the way split detect works to better match THE DOCUMENTATION that was annoying. Not sure if it's a real problem though or just an artifact of how I was testing.